Map updates and interactivity
Multiple interactivity features added, including click to expand events
on the left side and hover and click events on markers. Paths are drawn
on the map, statically for now. Clicking either a player's marker or a
game from the left pane opens the game view, where each player is
displayed in their own mini-map, along with their progress on the left.